Add APIs to test_com.android.sdkext module
This is used to verify in testing that new APIs added in a module
update are callable.

It isn't currently possible to build the test_framework-sdkextensions
java module from source, because of limitations in the hiddenapi
build tooling. Check in a prebuilt dexjar for now, that has the contents
of the original framework-sdkextensions.jar as well as the additional
Test class.

Bug: 149815613
Test: atest sdkextensions_e2e_test
Change-Id: I93ed41849868d27c0e3c4eb84c4402242235f0d4