Add new proto database and tool to manage it
This change adds the schema of the new extension database and the
skeleton of a tool to manage it.

Each extension version maps to set of module version requirements,
and a text representation of this proto database is checked in
(currently empty). The tool is currently only able to convert the
database to its binary representation (which will be used on-device to
derive what extension we have currently). Also add a test for the
tool/binary generation.

The tool will be extended in future CLs to be able to add/update/remove
SDKs, as well performing various consistency checks on the DB.

Bug: 173188089
Test: atest gen_sdk_test
Change-Id: I918170a851719bcc5bc70568f4705629b500735b