Update sdk derivation logic
Add logic to read the extension database and update the logic that
figures out the current SDK level based on the modules present.

For each extension (R, S etc), we find the highest version where all the
modules in that extension fulfil the requirements, and set the extension
sysprops accordingly.

This follows the design in go/mainline-sdk-versioning-impl.

Update the derive_sdk_test with a few more interesting cases (though no
multi-extension tests yet, that's coming next).

For the java tests, add a database that is compatible with the existing
test apexes. Future changes will add more apex versions into the mix to
make the tests more interesting.

Bug: 173188089
Test: atest --test-mapping packages/modules/SdkExtensions
Change-Id: I8b550f18c19330b65d250ffc6d7db43976cf6f80