Support custom output locations for derive_classpath
derive_classpath needs to be run during OTA to tell dexopt what the
new bootclasspath is. To avoid poluting the /data partition this adds
support for specifying a custom location to write the bootclasspath
environment variables to.

This is all run in a chroot which has the post-ota /system, /vendor,
/system_ext, /apex, etc partitions all mounted. The otapreopt system
uses these and the /data partition to compile everything using the new
bootclasspath. Since this chroot doesn't have an installd,
system_server or anything else otapreopt and otapreopt_chroot are
responsible for setting up enough of the system to run dex2oat and
need to do it in a way that doesn't impact the running system. For
derive_classpath it does that by having the new classpaths be written
into a tempfile it passes.

Test: manual OTA on blueline
Bug: 186432034
Change-Id: Ia310b5d4558f708e74f86ae56a4b33d80366534f