Explicitly disable usesNonSdkApi
As this app was recently made to compile against module_current, the
build system now automatically inserts usesNonSdkApi="true" into its
manifest. We do not want to be whitelisted for hiddenapi access in this
test app, so explicitly set it to false.

Bug: 189114287
Test: aapt list -a sdkextensions_e2e_test_app.apk and inspect
Change-Id: I296869381e29752f0f923131ae5a5749753847e5
Merged-In: I296869381e29752f0f923131ae5a5749753847e5