Add reserved_disk group in derive_classpath.rc.
The service is responsible for deriving BOOTCLASSPATH definition and
writing it to file on /data for init.rc to read during boot process.

This may not help some test setups, where the phone is filled to 100%
via adb root for example or at build time. However, it helps in a case
user or an app on the phone fills /data to 100%, while mainline update
is staged. This would guarantee that the file can be generated by
using the reserved 50mb of space on disk.

Bug: 195097326
Test: launch_cvd && atest CtsClasspathsTestCases
Ignore-AOSP-First: sc-dev change
Change-Id: Ia46ef4192c5ee1f44dd1c75d086baac9ce1de2bc
(cherry picked from commit e5834114cb5653f69f50be9883fc55839ad70a4f)