Write generated data to a temp file first, attempt #2.
The service is invoked manually during otapreopt with a custom output
location, where the file is a temporary file in memfs.

Write to a temp file only in the case when the service is invoked from
init.rc and the output is located in /data.

Bug: 195097326
Test: launch_cvd && atest CtsClasspathsTestCases
Ignore-AOSP-First: sc-dev fix
Change-Id: I28dc12ef9a0573c6e6c7dd415dc72c280f7fa9bf
(cherry picked from commit 26198b71800cefc4ee285a2cb808653c9a34ecea)