Update derive_sdk implementation
Consider modules from previous extensions part of the newer extensions.
This makes sure the right thing happens when e.g. an R modules adds an
S+ API (contributes to the S extensions), that module is required for
the S extensions API bump.

In effect, this caps the S extension so that it is never higher than the
R extension.

Bug: 206992877
Test: atest derive_sdk_test
Change-Id: I79b356f08ecd05df5a296c7411f77a36c3211099