Update version asserts in SdkExt tests
Because CTS is versioned with the dessert release, and devices can
be required to pass old CTS with new modules, we have a requirement
to be able to update the allowed extension versions when new ones
are added.

Make this convenient by refactoring the version checks to allow a set
of versions. At HEAD, this is always just the current version, but in
older CTS branches, this can be extended to allow more (newer versions)

Bug: 222049920
Test: atest -p packages/modules/SdkExtensions
Change-Id: I874fc1128e36e9fb5006cad938cf1bbc0f5f8d40