Add tool to convert API files to sdk-ext-info file
When importing changes to the sdk-extensions-info file after a new
sdkext version has been finalized, the format of the API files need to
be converted. Add a tool to do that.

The output must still be amended with module name and SDKs, and the
results reviewed. Shell scripts can help with the first part.

Test: manual: ./api-to-sdk-ext-info.py $(gettop)/prebuilts/sdk/extensions/2/public/api/framework-sdkextensions.txt
Change-Id: Iae38b37c62d2926227174260aa2d12fc770b96e0