Add AT_LEAST_BASE expectation
Allow tests to specify they are expecting the version to be at least
the base version of the dessert the device is currently running on.
For example, Android S shipped with extension version 1, so the base
version is 1.

This is useful for when tests are run in a setting where not all modules
have been updated to the latest version.

Bug: 244423777
Bug: 270925022
Test: atest SdkExtensionsTest
Change-Id: I502751219f40517ff13aee1d5c1750c3a8e64c5c