Change the expectations for testing on Go
Since go trains don't include all modules, a go device might not
necessarily see the dessert extension versions bumped even when all
the trains have been installed correctly. This does not apply for
preload trains (that are full). So when a device is a Go-device
with evidence of modules being side-loaded, only require the version
is at least the base version.

In the future, we may be able to improve this slightly to expect the
version to equal that of the preloaded train.

Bug: 244423777
Test: atest SdkExtensionsTest (on a go device with/without sideloads)
Change-Id: I3e62f38d9a7104cd54b25a5049c184806ee07123