Add standalone Tiramisu test
Rather than extend existing tests which are already quite complicated
this adds a standalone Tiramisu test that verifies that the mapping
from package to enum and the other Tiramisu related config is correct
without duplicating other testing.

This is intended to be an example of how to write a test for a new
dessert release. A follow up change will follow this pattern for
testing UpsideDownCake.

Ignore-AOSP-First: HEALTH_FITNESS and CONFIG_INFRASTRUCTURE are only internal
Bug: 222051472
Bug: 266528259
Test: atest derive_sdk_test
Change-Id: I85f44f880fceef4e566edb13d5af9ed938663d24