Replace `EXPECT_<X>` functions with macros for better errors
Previously, when an `EXPECT_<X>` function failed it would always report
the failure as within the function making them difficult to debug.
Using macros insteads means the errors are reported on the line where
the macro is used making debugging easier.

Ignore-AOSP-First: HEALTH_FITNESS and CONFIG_INFRASTRUCTURE are only internal
Bug: 266528259
Test: atest derive_sdk_test
Change-Id: If80404b95de55eec9c17f2cabd2e22e9efa6951d