Fix test crash on Go
Passing MATCH_ALL to getInstalledModules actually causes it to return
non-installed modules as well, which happens in the case of tzdata
where each version of tzdata is listed in the module metadata.

This causes a crash because we try to get the package info of a package
that doesn't exist. Fix this by not passing MATCH_ALL.

Bug: 274406002
Test: atest SdkExtensionsTest (on go, with 2023-05 installed)
Change-Id: Ida56d5386f098fc62190a1fca6c0235646968d0b