Run derive_sdk --dump from SdkExtensionsTest
It's often been difficult to debug why SdkExtensionsTest is failing,
as the sdk versions of all the apexes we depend on is difficult to
deduce from the test report.

Improve this by running the derive_sdk dump code before running
the test methods in SdkExtensionTest, and log the output to logcat.

Because we don't have permission to run the derive_sdk binary, we
include the derive_sdk code in the test and call it via JNI. While
this doesn't match the code that's on device, it doesn't really matter
since all it does is dump the contents of various protos. This also
gives us the ability to run newer dump code on older releases, in case
we need to tweak the debug info.

Bug: 277891386
Test: atest SdkExtensionsTest
Change-Id: I71c84d901d0097fd31d516ce7b90efdfae44f1d4