Add flags for M-2024-05 for participating modules.
Add the required flags indicating the FlaggedAPIs which need to be part
of the sdk extensions finalization for M-2024-05.
Participating modules -
1. Adservices
2. HealthFitness
3. Mediaprovider
4. Tethering

Changed the order of missions in reverse chronological order to ensure
the latest mission is visible on the top.
Ensured that all the flags for a module are sorted alphabetically.

Ignore-AOSP-First: This change is only needed in UMP for now.
Bug: N/A
Test: N/A
Change-Id: I8eb1ff030a748ec5b91798c30566b748c3e0290f