[DON'T BLOCK] Test ownership migration rules
This CL is created as a best effort to migrate test targets
to the new android ownership model. If you find incorrect or unnecessary
attribution in this CL, please create a separate CL to fix that.

For more details please refer to the link below,
<add g3 doc link>

Bug: 304529413
Test: N/A
Ignore-AOSP-First: The aosp migration is complete and this delta
Change-Id: I771eeaaddd21768888c2cd72bc611d01c4822538