SdkExtensionsTest: include DeriveSdk.dump in stack trace
Include the DeriveSdk.dump output in the stack trace of failing assert
of SdkExtensionsTest. This makes it immediately obvious which modules
that are keeping back the SDK extension version, which in turns means we
don't need to ask for a bug report or logcat log when we receive a test
failure bug.

Bug: 415761956
Test: atest CtsSdkExtensionsTestCases
Flag: EXEMPT adding more info when tests fail
Change-Id: I1b88c28f2278ff55426268f5945ca9431a303c1c