SdkExtensionsTest: verify that installed SdkExtensions and tests are compatible
Because SdkExtensionsTest verifies installed modules based on the
contents of extensions_db.textpb at the time the tests were built,
running an older version of the tests on a more recent version of the
installed SdkExtensions module is not a supported configuration. For
example, we could have just bumped the SDK extension version but forgot
to install the newer versions of the modules, and the old version of
SdkExtensions would not be able to detect this, happily claiming that
all tests pass.

Add an explicit check to SdkExtensionsTest to verify that the version of
the installed SdkExtensions module is the same or more recent than the
version of the tests.

Bug: 415761956
Test: atest CtsSdkExtensionsTestCases
Flag: EXEMPT adding instructions to unit test
Change-Id: I4cca365203c1a4dab3234c66f8a1e0b027135b2a